# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tenet Audit — EU AI Act Audit Log API",
    "description": "Tamper-evident AI call logging, risk classification (Article 6 / Annex III), incident reporting (Article 73) and auditor-ready evidence packs (Article 12).",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Tenant key (ta_live_…) or admin key (ta_admin_…)"
      }
    },
...
  },
  "paths": {
...
    "/v1/incident-report": {
      "post": {
        "tags": [
          "incidents"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "occurred_at",
                  "severity",
                  "title",
                  "description"
                ],
                "properties": {
                  "occurred_at": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "severity": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "serious"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "major"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "minor"
                        ]
                      }
                    ]
                  },
                  "title": {
                    "minLength": 1,
                    "maxLength": 300,
                    "type": "string"
                  },
                  "description": {
                    "minLength": 1,
                    "maxLength": 20000,
                    "type": "string"
                  },
                  "affected_users": {
                    "minimum": 0,
                    "type": "integer"
                  },
                  "related_log_ids": {
                    "maxItems": 500,
                    "type": "array",
                    "items": {
                      "format": "uuid",
                      "type": "string"
                    }
                  },
                  "remediation": {
                    "type": "object",
                    "properties": {
                      "actions": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "status": {
                        "type": "string"
                      },
                      "completed_at": {
                        "format": "date-time",
                        "type": "string"
                      }
                    }
                  },
                  "is_article_73": {
                    "type": "boolean"
                  },
                  "reported_to_authority": {
                    "type": "boolean"
                  },
                  "authority_ref": {
                    "maxLength": 500,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "required": [
                    "id",
                    "created_at",
                    "occurred_at",
                    "severity",
                    "is_article_73",
                    "title",
                    "description",
                    "affected_users",
                    "related_log_ids",
                    "remediation",
                    "reported_to_authority",
                    "authority_ref",
                    "status"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "occurred_at": {
                      "type": "string"
                    },
                    "severity": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "serious"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "major"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "minor"
                          ]
                        }
                      ]
                    },
                    "is_article_73": {
                      "type": "boolean"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "affected_users": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "related_log_ids": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "remediation": {},
                    "reported_to_authority": {
                      "type": "boolean"
                    },
                    "authority_ref": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "open"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "investigating"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "remediated"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "closed"
                          ]
                        }
                      ]
                    },
                    "article_73_note": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/incidents": {
      "get": {
        "tags": [
          "incidents"
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "serious"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "major"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "minor"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "severity",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "open"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "investigating"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "remediated"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "closed"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "format": "date-time",
              "type": "string"
            },
            "in": "query",
            "name": "from",
            "required": false
          },
          {
            "schema": {
              "format": "date-time",
              "type": "string"
            },
            "in": "query",
            "name": "to",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "additionalProperties": true,
                    "type": "object",
                    "required": [
                      "id",
                      "created_at",
                      "occurred_at",
                      "severity",
                      "is_article_73",
                      "title",
                      "description",
                      "affected_users",
                      "related_log_ids",
                      "remediation",
                      "reported_to_authority",
                      "authority_ref",
                      "status"
                    ],
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "created_at": {
                        "type": "string"
                      },
                      "occurred_at": {
                        "type": "string"
                      },
                      "severity": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "serious"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "major"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "minor"
                            ]
                          }
                        ]
                      },
                      "is_article_73": {
                        "type": "boolean"
                      },
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "affected_users": {
                        "anyOf": [
                          {
                            "type": "integer"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "related_log_ids": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "remediation": {},
                      "reported_to_authority": {
                        "type": "boolean"
                      },
                      "authority_ref": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "status": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "open"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "investigating"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "remediated"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "closed"
                            ]
                          }
                        ]
                      },
                      "article_73_note": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/incidents/{id}": {
      "patch": {
        "tags": [
          "incidents"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "open"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "investigating"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "remediated"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "closed"
                        ]
                      }
                    ]
                  },
                  "remediation": {
                    "type": "object",
                    "properties": {
                      "actions": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "status": {
                        "type": "string"
                      },
                      "completed_at": {
                        "format": "date-time",
                        "type": "string"
                      }
                    }
                  },
                  "reported_to_authority": {
                    "type": "boolean"
                  },
                  "authority_ref": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "affected_users": {
                    "minimum": 0,
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "required": [
                    "id",
                    "created_at",
                    "occurred_at",
                    "severity",
                    "is_article_73",
                    "title",
                    "description",
                    "affected_users",
                    "related_log_ids",
                    "remediation",
                    "reported_to_authority",
                    "authority_ref",
                    "status"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "occurred_at": {
                      "type": "string"
                    },
                    "severity": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "serious"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "major"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "minor"
                          ]
                        }
                      ]
                    },
                    "is_article_73": {
                      "type": "boolean"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "affected_users": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "related_log_ids": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "remediation": {},
                    "reported_to_authority": {
                      "type": "boolean"
                    },
                    "authority_ref": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "open"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "investigating"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "remediated"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "closed"
                          ]
                        }
                      ]
                    },
                    "article_73_note": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    }
  ],
...
}